+++ packages/documentation-examples/src/00 Chessboard/Knight.tsx
@@ -1,5 +1,5 @@
 import React, { useRef, useMemo } from 'react'
-import { useDragSource } from 'react-dnd'
+import { useDrag } from 'react-dnd'
 import ItemTypes from './ItemTypes'
 import knightImage from './knightImage'
 
@@ -21,7 +21,7 @@ function createKnightImage() {
 export const Knight: React.FC = () => {
 	const ref = useRef(null)
 	const dragPreview = useMemo(createKnightImage, [])
-	const { isDragging } = useDragSource(ref, ItemTypes.KNIGHT, {
+	const { isDragging } = useDrag(ref, ItemTypes.KNIGHT, {
 		beginDrag: () => ({}),
 		dragPreview,
 		collect: mon => ({
